Split the files so it makes it easier to manage
