+++ plugin/main_testing.go
@@ -1,0 +1,3 @@
+package main
+
+import "C"
